% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AttrGagePlan.R
\name{AttrGagePlan}
\alias{AttrGagePlan}
\title{AttrGagePlan: Attribute Agreement Analysis Plan}
\description{
Manage parameter configuration for attribute agreement analysis.
}
\section{Super class}{
\code{\link[iQualityR.core:IqrPlanBase]{iQualityR.core::IqrPlanBase}} -> \code{AttrGagePlan}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{appraisers}}{Vector of appraiser identifiers.}

\item{\code{samples}}{Vector of sample identifiers.}

\item{\code{standards}}{Vector of known reference standards for samples.}

\item{\code{categories}}{Vector of evaluation categories.}

\item{\code{category_order}}{Ordered vector of categories.}

\item{\code{scale_type}}{Character scalar: \code{"nominal"} or \code{"ordinal"}.}

\item{\code{trials}}{Integer scalar number of trials per appraiser/sample.}

\item{\code{comparison_mode}}{Character scalar: \code{"one_way"} or \code{"two_way"}.}

\item{\code{kappa_method}}{Character scalar: \code{"cohen"}, \code{"fleiss"}, or \code{"wagner"}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AttrGagePlan-new}{\code{AttrGagePlan$new()}}
\item \href{#method-AttrGagePlan-set_appraisers}{\code{AttrGagePlan$set_appraisers()}}
\item \href{#method-AttrGagePlan-set_samples}{\code{AttrGagePlan$set_samples()}}
\item \href{#method-AttrGagePlan-set_standards}{\code{AttrGagePlan$set_standards()}}
\item \href{#method-AttrGagePlan-set_categories}{\code{AttrGagePlan$set_categories()}}
\item \href{#method-AttrGagePlan-set_category_order}{\code{AttrGagePlan$set_category_order()}}
\item \href{#method-AttrGagePlan-set_scale_type}{\code{AttrGagePlan$set_scale_type()}}
\item \href{#method-AttrGagePlan-validate}{\code{AttrGagePlan$validate()}}
\item \href{#method-AttrGagePlan-generate_protocol}{\code{AttrGagePlan$generate_protocol()}}
\item \href{#method-AttrGagePlan-to_list}{\code{AttrGagePlan$to_list()}}
\item \href{#method-AttrGagePlan-clone}{\code{AttrGagePlan$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlanBase" data-id="set_criteria"><a href='../../iQualityR.core/html/IqrPlanBase.html#method-IqrPlanBase-set_criteria'><code>iQualityR.core::IqrPlanBase$set_criteria()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlanBase" data-id="set_meta"><a href='../../iQualityR.core/html/IqrPlanBase.html#method-IqrPlanBase-set_meta'><code>iQualityR.core::IqrPlanBase$set_meta()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttrGagePlan-new"></a>}}
\if{latex}{\out{\hypertarget{method-AttrGagePlan-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttrGagePlan$new(
  plan_name = "Attr_Gage_Study",
  objectives = "Attribute agreement analysis",
  appraisers = NULL,
  samples = NULL,
  standards = NULL,
  categories = NULL,
  category_order = NULL,
  scale_type = c("nominal", "ordinal"),
  trials = 2,
  comparison_mode = c("one_way", "two_way"),
  kappa_method = c("cohen", "fleiss", "wagner"),
  conf_level = 0.95
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plan_name}}{Character scalar project name.}

\item{\code{objectives}}{Character scalar study objectives.}

\item{\code{appraisers}}{Vector of appraiser identifiers.}

\item{\code{samples}}{Vector of sample identifiers.}

\item{\code{standards}}{Vector of known reference standards for samples.}

\item{\code{categories}}{Vector of evaluation categories.}

\item{\code{category_order}}{Ordered vector of categories.}

\item{\code{scale_type}}{Character scalar: \code{"nominal"} or \code{"ordinal"}.}

\item{\code{trials}}{Integer scalar number of trials per appraiser/sample.}

\item{\code{comparison_mode}}{Character scalar: \code{"one_way"} or \code{"two_way"}.}

\item{\code{kappa_method}}{Character scalar: \code{"cohen"}, \code{"fleiss"}, or \code{"wagner"}.}

\item{\code{conf_level}}{Confidence level (default 0.95).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttrGagePlan-set_appraisers"></a>}}
\if{latex}{\out{\hypertarget{method-AttrGagePlan-set_appraisers}{}}}
\subsection{Method \code{set_appraisers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttrGagePlan$set_appraisers(appraisers)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{appraisers}}{Vector of appraiser identifiers.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttrGagePlan-set_samples"></a>}}
\if{latex}{\out{\hypertarget{method-AttrGagePlan-set_samples}{}}}
\subsection{Method \code{set_samples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttrGagePlan$set_samples(samples)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{samples}}{Vector of sample identifiers.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttrGagePlan-set_standards"></a>}}
\if{latex}{\out{\hypertarget{method-AttrGagePlan-set_standards}{}}}
\subsection{Method \code{set_standards()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttrGagePlan$set_standards(standards)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{standards}}{Vector of known reference standards for samples.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttrGagePlan-set_categories"></a>}}
\if{latex}{\out{\hypertarget{method-AttrGagePlan-set_categories}{}}}
\subsection{Method \code{set_categories()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttrGagePlan$set_categories(categories)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{categories}}{Vector of evaluation categories.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttrGagePlan-set_category_order"></a>}}
\if{latex}{\out{\hypertarget{method-AttrGagePlan-set_category_order}{}}}
\subsection{Method \code{set_category_order()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttrGagePlan$set_category_order(category_order)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{category_order}}{Ordered vector of categories.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttrGagePlan-set_scale_type"></a>}}
\if{latex}{\out{\hypertarget{method-AttrGagePlan-set_scale_type}{}}}
\subsection{Method \code{set_scale_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttrGagePlan$set_scale_type(scale_type = c("nominal", "ordinal"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scale_type}}{Character scalar: \code{"nominal"} or \code{"ordinal"}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttrGagePlan-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AttrGagePlan-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttrGagePlan$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttrGagePlan-generate_protocol"></a>}}
\if{latex}{\out{\hypertarget{method-AttrGagePlan-generate_protocol}{}}}
\subsection{Method \code{generate_protocol()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttrGagePlan$generate_protocol()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttrGagePlan-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-AttrGagePlan-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttrGagePlan$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttrGagePlan-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AttrGagePlan-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttrGagePlan$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
