% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AttrGagePlotter.R
\name{AttrGagePlotter}
\alias{AttrGagePlotter}
\title{AttrGagePlotter}
\description{
Plotter for attribute agreement analysis.
}
\section{Super class}{
\code{\link[iQualityR.core:IqrPlotterBase]{iQualityR.core::IqrPlotterBase}} -> \code{AttrGagePlotter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AttrGagePlotter-new}{\code{AttrGagePlotter$new()}}
\item \href{#method-AttrGagePlotter-available_plots}{\code{AttrGagePlotter$available_plots()}}
\item \href{#method-AttrGagePlotter-render}{\code{AttrGagePlotter$render()}}
\item \href{#method-AttrGagePlotter-generate_plots}{\code{AttrGagePlotter$generate_plots()}}
\item \href{#method-AttrGagePlotter-plot_kappa_funnel}{\code{AttrGagePlotter$plot_kappa_funnel()}}
\item \href{#method-AttrGagePlotter-plot_kappa_benchmark}{\code{AttrGagePlotter$plot_kappa_benchmark()}}
\item \href{#method-AttrGagePlotter-plot_confusion_matrix}{\code{AttrGagePlotter$plot_confusion_matrix()}}
\item \href{#method-AttrGagePlotter-plot_pairwise_agreement}{\code{AttrGagePlotter$plot_pairwise_agreement()}}
\item \href{#method-AttrGagePlotter-plot_appraiser_standard}{\code{AttrGagePlotter$plot_appraiser_standard()}}
\item \href{#method-AttrGagePlotter-plot_response_agreement}{\code{AttrGagePlotter$plot_response_agreement()}}
\item \href{#method-AttrGagePlotter-plot_sample_disagreement}{\code{AttrGagePlotter$plot_sample_disagreement()}}
\item \href{#method-AttrGagePlotter-plot_kendall_ordinal}{\code{AttrGagePlotter$plot_kendall_ordinal()}}
\item \href{#method-AttrGagePlotter-plot_detection_ci}{\code{AttrGagePlotter$plot_detection_ci()}}
\item \href{#method-AttrGagePlotter-plot_detection_radar}{\code{AttrGagePlotter$plot_detection_radar()}}
\item \href{#method-AttrGagePlotter-plot_detection_metrics}{\code{AttrGagePlotter$plot_detection_metrics()}}
\item \href{#method-AttrGagePlotter-clone}{\code{AttrGagePlotter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttrGagePlotter-new"></a>}}
\if{latex}{\out{\hypertarget{method-AttrGagePlotter-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttrGagePlotter$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttrGagePlotter-available_plots"></a>}}
\if{latex}{\out{\hypertarget{method-AttrGagePlotter-available_plots}{}}}
\subsection{Method \code{available_plots()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttrGagePlotter$available_plots(results = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{Results list from \link{AttrGageTask}$build_results() or analyzer raw output.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttrGagePlotter-render"></a>}}
\if{latex}{\out{\hypertarget{method-AttrGagePlotter-render}{}}}
\subsection{Method \code{render()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttrGagePlotter$render(results, theme_obj = NULL, type = "summary", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{Results list from \link{AttrGageTask}$build_results() or analyzer raw output.}

\item{\code{theme_obj}}{Theme object produced by IqrTheme.}

\item{\code{type}}{Plot type: \code{"summary"} or \code{"list"}.}

\item{\code{...}}{Additional arguments passed to downstream methods.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttrGagePlotter-generate_plots"></a>}}
\if{latex}{\out{\hypertarget{method-AttrGagePlotter-generate_plots}{}}}
\subsection{Method \code{generate_plots()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttrGagePlotter$generate_plots(results, theme_obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{Results list from \link{AttrGageTask}$build_results() or analyzer raw output.}

\item{\code{theme_obj}}{Theme object produced by IqrTheme.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttrGagePlotter-plot_kappa_funnel"></a>}}
\if{latex}{\out{\hypertarget{method-AttrGagePlotter-plot_kappa_funnel}{}}}
\subsection{Method \code{plot_kappa_funnel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttrGagePlotter$plot_kappa_funnel(results, theme_obj = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{Results list from \link{AttrGageTask}$build_results() or analyzer raw output.}

\item{\code{theme_obj}}{Theme object produced by IqrTheme.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttrGagePlotter-plot_kappa_benchmark"></a>}}
\if{latex}{\out{\hypertarget{method-AttrGagePlotter-plot_kappa_benchmark}{}}}
\subsection{Method \code{plot_kappa_benchmark()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttrGagePlotter$plot_kappa_benchmark(results, theme_obj = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{Results list from \link{AttrGageTask}$build_results() or analyzer raw output.}

\item{\code{theme_obj}}{Theme object produced by IqrTheme.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttrGagePlotter-plot_confusion_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-AttrGagePlotter-plot_confusion_matrix}{}}}
\subsection{Method \code{plot_confusion_matrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttrGagePlotter$plot_confusion_matrix(results, theme_obj = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{Results list from \link{AttrGageTask}$build_results() or analyzer raw output.}

\item{\code{theme_obj}}{Theme object produced by IqrTheme.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttrGagePlotter-plot_pairwise_agreement"></a>}}
\if{latex}{\out{\hypertarget{method-AttrGagePlotter-plot_pairwise_agreement}{}}}
\subsection{Method \code{plot_pairwise_agreement()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttrGagePlotter$plot_pairwise_agreement(results, theme_obj = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{Results list from \link{AttrGageTask}$build_results() or analyzer raw output.}

\item{\code{theme_obj}}{Theme object produced by IqrTheme.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttrGagePlotter-plot_appraiser_standard"></a>}}
\if{latex}{\out{\hypertarget{method-AttrGagePlotter-plot_appraiser_standard}{}}}
\subsection{Method \code{plot_appraiser_standard()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttrGagePlotter$plot_appraiser_standard(results, theme_obj = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{Results list from \link{AttrGageTask}$build_results() or analyzer raw output.}

\item{\code{theme_obj}}{Theme object produced by IqrTheme.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttrGagePlotter-plot_response_agreement"></a>}}
\if{latex}{\out{\hypertarget{method-AttrGagePlotter-plot_response_agreement}{}}}
\subsection{Method \code{plot_response_agreement()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttrGagePlotter$plot_response_agreement(results, theme_obj = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{Results list from \link{AttrGageTask}$build_results() or analyzer raw output.}

\item{\code{theme_obj}}{Theme object produced by IqrTheme.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttrGagePlotter-plot_sample_disagreement"></a>}}
\if{latex}{\out{\hypertarget{method-AttrGagePlotter-plot_sample_disagreement}{}}}
\subsection{Method \code{plot_sample_disagreement()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttrGagePlotter$plot_sample_disagreement(results, theme_obj = NULL, top_n = 12)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{Results list from \link{AttrGageTask}$build_results() or analyzer raw output.}

\item{\code{theme_obj}}{Theme object produced by IqrTheme.}

\item{\code{top_n}}{Integer scalar number of samples to display in disagreement plot.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttrGagePlotter-plot_kendall_ordinal"></a>}}
\if{latex}{\out{\hypertarget{method-AttrGagePlotter-plot_kendall_ordinal}{}}}
\subsection{Method \code{plot_kendall_ordinal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttrGagePlotter$plot_kendall_ordinal(results, theme_obj = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{Results list from \link{AttrGageTask}$build_results() or analyzer raw output.}

\item{\code{theme_obj}}{Theme object produced by IqrTheme.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttrGagePlotter-plot_detection_ci"></a>}}
\if{latex}{\out{\hypertarget{method-AttrGagePlotter-plot_detection_ci}{}}}
\subsection{Method \code{plot_detection_ci()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttrGagePlotter$plot_detection_ci(results, theme_obj = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{Results list from \link{AttrGageTask}$build_results() or analyzer raw output.}

\item{\code{theme_obj}}{Theme object produced by IqrTheme.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttrGagePlotter-plot_detection_radar"></a>}}
\if{latex}{\out{\hypertarget{method-AttrGagePlotter-plot_detection_radar}{}}}
\subsection{Method \code{plot_detection_radar()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttrGagePlotter$plot_detection_radar(results, theme_obj = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{Results list from \link{AttrGageTask}$build_results() or analyzer raw output.}

\item{\code{theme_obj}}{Theme object produced by IqrTheme.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttrGagePlotter-plot_detection_metrics"></a>}}
\if{latex}{\out{\hypertarget{method-AttrGagePlotter-plot_detection_metrics}{}}}
\subsection{Method \code{plot_detection_metrics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttrGagePlotter$plot_detection_metrics(results, theme_obj = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{Results list from \link{AttrGageTask}$build_results() or analyzer raw output.}

\item{\code{theme_obj}}{Theme object produced by IqrTheme.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttrGagePlotter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AttrGagePlotter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttrGagePlotter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
