% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetectionAnalyzer.R
\name{DetectionAnalyzer}
\alias{DetectionAnalyzer}
\title{DetectionAnalyzer: Detection Rate and False Negative Rate Analysis}
\description{
Statistical analysis of detection rate and false negative rate using binomial distribution
}
\section{Super class}{
\code{\link[iQualityR.core:IqrAnalyzerBase]{iQualityR.core::IqrAnalyzerBase}} -> \code{DetectionAnalyzer}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{detection_rate}}{Proportion of positive reference samples correctly detected.}

\item{\code{false_negative_rate}}{Proportion of positive reference samples missed.}

\item{\code{binomial_result}}{Clopper-Pearson confidence interval result for the detection rate.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DetectionAnalyzer-new}{\code{DetectionAnalyzer$new()}}
\item \href{#method-DetectionAnalyzer-run}{\code{DetectionAnalyzer$run()}}
\item \href{#method-DetectionAnalyzer-analyze_detection}{\code{DetectionAnalyzer$analyze_detection()}}
\item \href{#method-DetectionAnalyzer-binomial_ci}{\code{DetectionAnalyzer$binomial_ci()}}
\item \href{#method-DetectionAnalyzer-detection_test}{\code{DetectionAnalyzer$detection_test()}}
\item \href{#method-DetectionAnalyzer-summary_table}{\code{DetectionAnalyzer$summary_table()}}
\item \href{#method-DetectionAnalyzer-clone}{\code{DetectionAnalyzer$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="get_results"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-get_results'><code>iQualityR.core::IqrAnalyzerBase$get_results()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="reset"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-reset'><code>iQualityR.core::IqrAnalyzerBase$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="set_datatable"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-set_datatable'><code>iQualityR.core::IqrAnalyzerBase$set_datatable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="set_diagnostic"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-set_diagnostic'><code>iQualityR.core::IqrAnalyzerBase$set_diagnostic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="set_raw_output"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-set_raw_output'><code>iQualityR.core::IqrAnalyzerBase$set_raw_output()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="set_statistic"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-set_statistic'><code>iQualityR.core::IqrAnalyzerBase$set_statistic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="setup"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-setup'><code>iQualityR.core::IqrAnalyzerBase$setup()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DetectionAnalyzer-new"></a>}}
\if{latex}{\out{\hypertarget{method-DetectionAnalyzer-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DetectionAnalyzer$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DetectionAnalyzer-run"></a>}}
\if{latex}{\out{\hypertarget{method-DetectionAnalyzer-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DetectionAnalyzer$run(data, plan = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data frame with reference and test columns.}

\item{\code{plan}}{Optional plan object or list of parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DetectionAnalyzer-analyze_detection"></a>}}
\if{latex}{\out{\hypertarget{method-DetectionAnalyzer-analyze_detection}{}}}
\subsection{Method \code{analyze_detection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DetectionAnalyzer$analyze_detection(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data frame with reference and test columns.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DetectionAnalyzer-binomial_ci"></a>}}
\if{latex}{\out{\hypertarget{method-DetectionAnalyzer-binomial_ci}{}}}
\subsection{Method \code{binomial_ci()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DetectionAnalyzer$binomial_ci(x, n, conf_level = 0.95)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Number of successes (e.g. true positives).}

\item{\code{n}}{Number of trials (e.g. total positive reference samples).}

\item{\code{conf_level}}{Confidence level (default 0.95).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DetectionAnalyzer-detection_test"></a>}}
\if{latex}{\out{\hypertarget{method-DetectionAnalyzer-detection_test}{}}}
\subsection{Method \code{detection_test()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DetectionAnalyzer$detection_test(null_rate = 0.8, alternative = "greater")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{null_rate}}{Null hypothesis rate for the binomial test.}

\item{\code{alternative}}{Alternative hypothesis: \code{"greater"}, \code{"less"}, or \code{"two.sided"}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DetectionAnalyzer-summary_table"></a>}}
\if{latex}{\out{\hypertarget{method-DetectionAnalyzer-summary_table}{}}}
\subsection{Method \code{summary_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DetectionAnalyzer$summary_table()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DetectionAnalyzer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DetectionAnalyzer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DetectionAnalyzer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
