% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa_functions.R
\name{gage_rr_xbar_r}
\alias{gage_rr_xbar_r}
\title{MSA Gage R&R Study (Xbar-R Method)}
\usage{
gage_rr_xbar_r(
  data,
  measurement_col = "MeasurementValue",
  part_col = "Part",
  operator_col = "Operator",
  theme = "academic",
  sigma = 6,
  tolerance = NULL,
  lsl = NULL,
  usl = NULL
)
}
\arguments{
\item{data}{Measurement data (data.frame).}

\item{measurement_col}{Measurement value column name.}

\item{part_col}{Part column name.}

\item{operator_col}{Operator column name.}

\item{theme}{Theme name.}

\item{sigma}{Process variation multiplier.}

\item{tolerance}{Study tolerance width used for Percent Tolerance.}

\item{lsl}{Lower specification limit.}

\item{usl}{Upper specification limit.}
}
\value{
An \link{MsaTask} object containing analysis results, plots, and report methods.
}
\description{
Perform a Gage Repeatability and Reproducibility study using the Xbar-R method.
}
