% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa_functions.R
\name{infer_msa_plan}
\alias{infer_msa_plan}
\title{Infer MSA Plan from Data}
\usage{
infer_msa_plan(
  data,
  measurement_col = "MeasurementValue",
  part_col = "Part",
  operator_col = "Operator"
)
}
\arguments{
\item{data}{Measurement data.}

\item{measurement_col}{Measurement value column name.}

\item{part_col}{Part column name.}

\item{operator_col}{Operator column name.}
}
\value{
An MSAPlan object.
}
\description{
Infer MSA plan parameters from existing measurement data.
}
