% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr_gage_functions.R
\name{iqr_cohen_kappa}
\alias{iqr_cohen_kappa}
\title{Cohen's Kappa analysis shortcut entry}
\usage{
iqr_cohen_kappa(
  data,
  eval1_col = "Evaluator1",
  eval2_col = "Evaluator2",
  theme = "academic"
)
}
\arguments{
\item{data}{Data frame, must contain two evaluation result columns}

\item{eval1_col}{First evaluator column name}

\item{eval2_col}{Second evaluator column name}

\item{theme}{Theme name (default "academic").}
}
\description{
Cohen's Kappa analysis shortcut entry
}
