% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr_gage_functions.R
\name{iqr_wagner}
\alias{iqr_wagner}
\title{Wagner method analysis shortcut entry}
\usage{
iqr_wagner(
  data,
  eval1_col = "Evaluator1",
  eval2_col = "Evaluator2",
  theme = "academic"
)
}
\arguments{
\item{data}{Data frame}

\item{eval1_col}{First evaluator column name}

\item{eval2_col}{Second evaluator column name}

\item{theme}{Theme name (default "academic").}
}
\description{
Wagner method analysis shortcut entry
}
