% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{iQualityR.msa-package}
\alias{iQualityR.msa}
\alias{iQualityR.msa-package}
\title{iQualityR.msa}
\description{
Measurement system analysis tools for the iQualityR ecosystem.
}
\author{
\strong{Maintainer}: iQualityR Team \email{iQualityR@example.com}

}
\keyword{internal}
