% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KappaAnalyzer.R
\name{KappaAnalyzer}
\alias{KappaAnalyzer}
\title{KappaAnalyzer}
\description{
Calculates Cohen/Fleiss kappa and Minitab-style agreement tables.
}
\section{Super class}{
\code{\link[iQualityR.core:IqrAnalyzerBase]{iQualityR.core::IqrAnalyzerBase}} -> \code{KappaAnalyzer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KappaAnalyzer-new}{\code{KappaAnalyzer$new()}}
\item \href{#method-KappaAnalyzer-run}{\code{KappaAnalyzer$run()}}
\item \href{#method-KappaAnalyzer-one_way_analysis}{\code{KappaAnalyzer$one_way_analysis()}}
\item \href{#method-KappaAnalyzer-cohen_kappa}{\code{KappaAnalyzer$cohen_kappa()}}
\item \href{#method-KappaAnalyzer-fleiss_kappa}{\code{KappaAnalyzer$fleiss_kappa()}}
\item \href{#method-KappaAnalyzer-two_way_wagner}{\code{KappaAnalyzer$two_way_wagner()}}
\item \href{#method-KappaAnalyzer-interpret_kappa}{\code{KappaAnalyzer$interpret_kappa()}}
\item \href{#method-KappaAnalyzer-clone}{\code{KappaAnalyzer$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="get_results"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-get_results'><code>iQualityR.core::IqrAnalyzerBase$get_results()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="reset"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-reset'><code>iQualityR.core::IqrAnalyzerBase$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="set_datatable"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-set_datatable'><code>iQualityR.core::IqrAnalyzerBase$set_datatable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="set_diagnostic"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-set_diagnostic'><code>iQualityR.core::IqrAnalyzerBase$set_diagnostic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="set_raw_output"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-set_raw_output'><code>iQualityR.core::IqrAnalyzerBase$set_raw_output()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="set_statistic"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-set_statistic'><code>iQualityR.core::IqrAnalyzerBase$set_statistic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="setup"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-setup'><code>iQualityR.core::IqrAnalyzerBase$setup()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KappaAnalyzer-new"></a>}}
\if{latex}{\out{\hypertarget{method-KappaAnalyzer-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KappaAnalyzer$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KappaAnalyzer-run"></a>}}
\if{latex}{\out{\hypertarget{method-KappaAnalyzer-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KappaAnalyzer$run(data, plan = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data frame of attribute ratings.}

\item{\code{plan}}{Optional plan object or list of parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KappaAnalyzer-one_way_analysis"></a>}}
\if{latex}{\out{\hypertarget{method-KappaAnalyzer-one_way_analysis}{}}}
\subsection{Method \code{one_way_analysis()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KappaAnalyzer$one_way_analysis(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data frame of attribute ratings.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KappaAnalyzer-cohen_kappa"></a>}}
\if{latex}{\out{\hypertarget{method-KappaAnalyzer-cohen_kappa}{}}}
\subsection{Method \code{cohen_kappa()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KappaAnalyzer$cohen_kappa(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data frame of attribute ratings.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KappaAnalyzer-fleiss_kappa"></a>}}
\if{latex}{\out{\hypertarget{method-KappaAnalyzer-fleiss_kappa}{}}}
\subsection{Method \code{fleiss_kappa()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KappaAnalyzer$fleiss_kappa(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data frame of attribute ratings.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KappaAnalyzer-two_way_wagner"></a>}}
\if{latex}{\out{\hypertarget{method-KappaAnalyzer-two_way_wagner}{}}}
\subsection{Method \code{two_way_wagner()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KappaAnalyzer$two_way_wagner(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data frame of attribute ratings.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KappaAnalyzer-interpret_kappa"></a>}}
\if{latex}{\out{\hypertarget{method-KappaAnalyzer-interpret_kappa}{}}}
\subsection{Method \code{interpret_kappa()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KappaAnalyzer$interpret_kappa(kappa)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{kappa}}{Numeric scalar kappa value to interpret.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KappaAnalyzer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KappaAnalyzer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KappaAnalyzer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
