% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeteringStudy.R
\name{MeteringStudy}
\alias{MeteringStudy}
\title{MeteringStudy}
\description{
Backward-compatible alias for \code{MsaTask}. New code should use
\code{MsaTask}; this wrapper preserves existing user code that calls
\code{MeteringStudy$new()}.
}
\section{Super classes}{
\code{\link[iQualityR.core:IqrTaskBase]{iQualityR.core::IqrTaskBase}} -> \code{\link[iQualityR.msa:MsaTask]{iQualityR.msa::MsaTask}} -> \code{MeteringStudy}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MeteringStudy-clone}{\code{MeteringStudy$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="iQualityR.msa" data-topic="MsaTask" data-id="build_excel_sheets"><a href='../../iQualityR.msa/html/MsaTask.html#method-MsaTask-build_excel_sheets'><code>iQualityR.msa::MsaTask$build_excel_sheets()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.msa" data-topic="MsaTask" data-id="build_report_results"><a href='../../iQualityR.msa/html/MsaTask.html#method-MsaTask-build_report_results'><code>iQualityR.msa::MsaTask$build_report_results()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.msa" data-topic="MsaTask" data-id="compute"><a href='../../iQualityR.msa/html/MsaTask.html#method-MsaTask-compute'><code>iQualityR.msa::MsaTask$compute()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.msa" data-topic="MsaTask" data-id="generate_plots"><a href='../../iQualityR.msa/html/MsaTask.html#method-MsaTask-generate_plots'><code>iQualityR.msa::MsaTask$generate_plots()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.msa" data-topic="MsaTask" data-id="initialize"><a href='../../iQualityR.msa/html/MsaTask.html#method-MsaTask-initialize'><code>iQualityR.msa::MsaTask$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.msa" data-topic="MsaTask" data-id="plot"><a href='../../iQualityR.msa/html/MsaTask.html#method-MsaTask-plot'><code>iQualityR.msa::MsaTask$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.msa" data-topic="MsaTask" data-id="report"><a href='../../iQualityR.msa/html/MsaTask.html#method-MsaTask-report'><code>iQualityR.msa::MsaTask$report()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.msa" data-topic="MsaTask" data-id="summary"><a href='../../iQualityR.msa/html/MsaTask.html#method-MsaTask-summary'><code>iQualityR.msa::MsaTask$summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.msa" data-topic="MsaTask" data-id="to_excel"><a href='../../iQualityR.msa/html/MsaTask.html#method-MsaTask-to_excel'><code>iQualityR.msa::MsaTask$to_excel()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeteringStudy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MeteringStudy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeteringStudy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
